# store-service/cmd/docs/swagger.yaml
definitions:
...
  product.NewProduct:
    properties:
      product-brand:
        type: string
      product-name:
        type: string
      product-price:
        type: number
      product-stock:
        type: integer
      product_image:
        type: string
    required:
    - product-brand
    - product-name
    - product-price
    - product-stock
    type: object
  product.Product:
    properties:
      id:
        type: integer
      product_image:
        type: string
      product_name:
        type: string
      product_price:
        type: number
      product_price_full_thb:
        type: number
      product_price_thb:
        type: number
    type: object
info:
  contact: {}
paths:
...
  /api/v1/product:
    get:
      consumes:
      - application/json
      description: Search for products with optional filtering
      parameters:
      - description: Search keyword
        in: query
        name: q
        type: string
      - default: "30"
        description: Number of items per page
        in: query
        name: limit
        type: string
      - default: "0"
        description: Offset for pagination
        in: query
        name: offset
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/product.Product'
            type: array
        "500":
          description: Internal Server Error
      summary: Search products
      tags:
      - product
    post:
      consumes:
      - application/json
      description: Create a new product
      parameters:
      - description: New product data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/product.NewProduct'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create new product
      tags:
      - product
  /api/v1/product/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.Product'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
      summary: Get product by ID
      tags:
      - product
swagger: "2.0"